+++ src/primitive.h
@@ -53,7 +53,7 @@ public:
 	virtual Primitive* boundary()=0;
 	virtual Primitive* copy()=0;
 	virtual Primitive* triangulate()=0;
-	virtual Primitive* simplify(int)=0;
+	virtual Primitive* simplify(decimal)=0;
 	virtual void transform(TransformMatrix*)=0;
 	virtual bool isEmpty()=0;
 	virtual bool isFullyDimentional()=0;
